refactor: 채팅방 메세지 컴포넌트 분리 및 로컬스토리지 토큰 관리 코드 유틸함수로 리팩토링
