+++ frontend/src/libs/constants/keys.ts
@@ -1,0 +1,5 @@
+const storageKeys = {
+  ACCESS_TOKEN: "accessToken",
+} as const;
+
+export { storageKeys };
